requirements.txt
# Core Python dependencies
python-dotenv         # Load environment variables
requests              # HTTP requests for API calls

# Machine Learning / LLM
openai
sentence_transfromers

# LangChain & LangGraph
langchain            # LLM orchestration and utilities
langgraph            # Graph-based AI pipelines
langchain_openai
langchain-text-splitters

# Simulation
pyglet

# Date/time handling
python-dateutil==2.8.2        # Parsing timestamps from telemetry